worked on linear hash functions.
hash put, get, and split works OK i think.

Got some problems with BucketIter. It may be wise to pass it as an argument
instead of as a return value from Bucket::iter().

Got some problems with bucket chunk head linkage. For some reason, the
links cannot be read from file. This will need to be corrected.